+++ package.json
@@ -1,32 +1,38 @@
 {
-    "devDependencies": {
-        "babel-core": "^6.26.3",
-        "babel-loader": "^7.1.4",
-        "babel-preset-env": "^1.7.0",
-        "babel-preset-es2015": "^6.24.1",
-        "babel-preset-react": "^6.24.1",
-        "body-parser": "^1.18.3",
-        "express": "^4.16.4",
-        "graphql": "^14.1.1",
-        "mongodb": "^3.1.13",
-        "nodemon": "^1.18.10",
-        "prop-types": "^15.7.2",
-        "webpack": "^4.15.1",
-        "webpack-cli": "^3.1.1",
-        "yarn": "^1.13.0"
-    },
-    "name": "question",
-    "version": "1.0.0",
-    "main": "service.js",
-    "scripts": {
-        "compile": "babel src --presets react,es2015 --out-dir static",
-        "watch": "babel src --presets react,es2015 --out-dir static --watch",
-        "dev": "node service/service.js",
-        "start": "nodemon -w service service/service.js",
-        "build": "webpack --mode none",
-        "buildproduction": "webpack --mode production"
-    },
-    "author": "",
-    "license": "ISC",
-    "description": ""
+  "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.4",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "body-parser": "^1.18.3",
+    "express": "^4.16.4",
+    "graphql": "^14.1.1",
+    "mongodb": "^3.1.13",
+    "nodemon": "^1.18.10",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
+    "webpack": "^4.15.1",
+    "webpack-cli": "^3.1.1",
+    "webpack-dev-server": "^3.2.1",
+    "whatwg-fetch": "^3.0.0",
+    "yarn": "^1.13.0"
+  },
+  "name": "question",
+  "version": "1.0.0",
+  "main": "service.js",
+  "scripts": {
+    "oldcompile": "babel src --presets react,es2015 --out-dir static",
+    "oldwatch": "babel src --presets react,es2015 --out-dir static --watch",
+    "dev": "node service/service.js",
+    "start": "nodemon -w service service/service.js",
+    "build": "webpack --mode development",
+    "buildproduction": "webpack --mode production",
+    "watch": "webpack --mode none --watch"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": ""
 }
