+++ src/IssueFilter.jsx
@@ -1,0 +1,7 @@
+import react from 'react';
+
+export default class IssueFilter extends React.Component {
+	render() {
+		return <div>Filter</div>
+	}
+}
